+++ R/dof.R
@@ -133,6 +133,9 @@ dof <- degrees_of_freedom
   if (is.null(method)) {
     return(TRUE)
   }
+  if (!insight::model_info(model)$is_mixed) {
+    return(FALSE)
+  }
   method <- tolower(method)
   if (!(method %in% c("analytical", "any", "fit", "satterthwaite", "betwithin", "kenward", "kr", "nokr", "wald", "ml1"))) {
     warning("'df_method' must be one of 'wald', 'kenward', 'satterthwaite', 'betwithin' or ' ml1'. Using 'wald' now.", call. = FALSE)