+++ R/model_parameters.R
@@ -13,12 +13,12 @@
 #'  \item{\link[=model_parameters.stanreg]{Bayesian models} (\pkg{rstanarm}, \pkg{brms}, \pkg{MCMCglmm})}
 #'  \item{\link[=model_parameters.principal]{PCA and FA} (\pkg{psych})}
 #'  \item{\link[=model_parameters.lavaan]{CFA and SEM} (\pkg{lavaan}, \pkg{blavaan})}
-#'  \item{\link[=model_parameters.kmeans]{Cluster models (k-means, ...)}}
+#'  \item{\link[=model_parameters.kmeans]{Cluster models} (k-means, ...)}
 #'  \item{\link[=model_parameters.rma]{Meta-Analysis via linear (mixed) models} (\code{rma})}
 #'  }
 #'
 #' @param model Statistical Model.
-#' @param ... Arguments passed to or from other methods.
+#' @param ... Arguments passed to or from other methods. Non-documented arguments are \code{digits}, \code{p_digits} and \code{ci_digits} to set the number of digits for the output. See 'Examples' in \code{\link{model_parameters.default}}.
 #'
 #' @seealso \code{\link[=standardize_names]{standardize_names()}} to rename
 #'   columns into a consistent, standardized naming scheme.
@@ -29,9 +29,6 @@
 #'   In case of \code{standardize = "refit"}, the data used to fit the model
 #'   will be standardized and the model is completely refitted. In such cases,
 #'   standard errors and confidence intervals refer to the standardized coefficient.
-#'   For other methods of standardizing, unstandardized and standardized coefficients
-#'   are shown, where standard errors and confidence intervals relate to the
-#'   unstandardized coefficients.
 #'
 #' @return A data frame of indices related to the model's parameters.
 #' @export