+++ R/random_parameters.R
@@ -2,13 +2,13 @@
 #' @name random_parameters
 #'
 #' @description This function extracts the different variance components of a
-#'   mixed model as well as further information related to random effects (ICC, R2)
-#'   and returns the result as a data frame.
+#'   mixed model and returns the result as a data frame.
 #'
 #' @param model A mixed effects model.
 #'
-#' @return A data frame with random effects statistics, including number of levels per
-#'   random effect group, as well as complete observations in the model.
+#' @return A data frame with random effects statistics for the variance components,
+#'   including number of levels per random effect group, as well as complete
+#'   observations in the model.
 #'
 #' @details The variance components are obtained from \code{\link[insight]{get_variance}}
 #'   and are denoted as following:
@@ -17,16 +17,18 @@
 #'     is the sum of the distribution-specific variance and the variance due to additive dispersion.
 #'     It indicates the \emph{within-group variance}.
 #'   }
-#'   \subsection{Between-group (or random intercept) variance}{
+#'   \subsection{Between-group random intercept variance}{
 #'     The random intercept variance, or \emph{between-group} variance
-#'     (\ifelse{html}{\out{&tau;<sub>00</sub>}}{\eqn{\tau_{00}}}),
+#'     for the intercept (\ifelse{html}{\out{&tau;<sub>00</sub>}}{\eqn{\tau_{00}}}),
 #'     is obtained from \code{VarCorr()}. It indicates how much groups
 #'     or subjects differ from each other.
 #'   }
-#'   \subsection{Random slope variance}{
-#'     The random slope variance (\ifelse{html}{\out{&tau;<sub>11</sub>}}{\eqn{\tau_{11}}})
+#'   \subsection{Between-group random slope variance}{
+#'     The random slope variance, or \emph{between-group} variance
+#'     for the slopes (\ifelse{html}{\out{&tau;<sub>11</sub>}}{\eqn{\tau_{11}}})
 #'     is obtained from \code{VarCorr()}. This measure is only available
-#'     for mixed models with random slopes.
+#'     for mixed models with random slopes. It indicates how much groups
+#'     or subjects differ from each other according to their slopes.
 #'   }
 #'   \subsection{Random slope-intercept correlation}{
 #'     The random slope-intercept correlation
@@ -34,8 +36,6 @@
 #'     is obtained from \code{VarCorr()}. This measure is only available
 #'     for mixed models with random intercepts and slopes.
 #'   }
-#'  For details regarding the calculation of R2 and ICC, see \code{\link[performance]{r2_nakagawa}}
-#'  and \code{\link[performance]{icc}}.
 #'
 #' @examples
 #' if (require("lme4")) {
@@ -68,8 +68,6 @@ random_parameters <- function(model) {
   model_re <- insight::find_random(model, split_nested = FALSE, flatten = TRUE)
   model_rs <- unlist(insight::find_random_slopes(model))
 
-  new_line <- FALSE
-
   if (length(re_variances) && sum(!is.na(re_variances)) > 0 && !is.null(re_variances)) {
     # Residual Variance (Sigma^2)
     out$Sigma2 <- re_variances$var.residual
@@ -81,8 +79,6 @@ random_parameters <- function(model) {
 
     # Random Slope Variance
     if (!.is_empty_object(re_variances$var.slope) && !.is_empty_object(model_rs)) {
-      out <- c(out, as.list(NA))
-      new_line <- TRUE
       var_slope <- as.list(re_variances$var.slope)
       names(var_slope) <- paste0("tau11_", names(re_variances$var.slope))
       out <- c(out, var_slope)
@@ -90,24 +86,10 @@ random_parameters <- function(model) {
 
     # Slope-Intercept Correlation
     if (!.is_empty_object(re_variances$cor.slope_intercept) && !.is_empty_object(model_rs)) {
-      if (!new_line) {
-        out <- c(out, as.list(NA))
-        new_line <- TRUE
-      }
       cor_slope_intercept <- as.list(re_variances$cor.slope_intercept)
       names(cor_slope_intercept) <- paste0("rho01_", model_re, ".", model_rs)
       out <- c(out, cor_slope_intercept)
     }
-
-    out <- c(out, as.list(NA))
-
-    # ICC & R2
-    out$R2_marginal <- re_variances$var.fixed / (re_variances$var.fixed + re_variances$var.residual)
-    if (!.is_empty_object(re_variances$var.random) && !is.na(re_variances$var.random)) {
-      out$R2_conditional <- (re_variances$var.fixed + re_variances$var.random) / (re_variances$var.fixed + re_variances$var.random + re_variances$var.residual)
-      out$ICC <- re_variances$var.random / (re_variances$var.random + re_variances$var.residual)
-    }
-    out <- c(out, as.list(NA))
   }
 
   # Number of levels per random-effect groups
@@ -138,18 +120,30 @@ random_parameters <- function(model) {
   out$Term[out$Component == "rho01"] <- gsub("^rho01_(.*)", "\\1", out$Description[out$Component == "rho01"])
 
   # renaming
-  out$Description[out$Description == "Sigma2"] <- "Within-Group Variance"
-  out$Description <- gsub("^tau00_(.*)", "Between-Group Variance", out$Description)
-  out$Description <- gsub("^tau11_(.*)", "Random Slope Variance", out$Description)
-  out$Description <- gsub("^rho01_(.*)", "Slope-Intercept Correlation", out$Description)
+  out$Type <- ""
 
+  # Within-Group Variance
+  out$Type[out$Description == "Sigma2"] <- ""
+  out$Description[out$Description == "Sigma2"] <- "Within-Group Variance"
+
+  # Between-Group Variance
+  out$Type[grepl("^tau00_", out$Description)] <- "Random Intercept"
+  out$Description <- gsub("^tau00_(.*)", "Between-Group Variance", out$Description)
+  out$Type[grepl("^tau11_", out$Description)] <- "Random Slope"
+  out$Description <- gsub("^tau11_(.*)", "Between-Group Variance", out$Description)
+
+  # correlations
+  out$Type[grepl("^rho01_", out$Description)] <- ""
+  out$Description <- gsub("^rho01_(.*)", "Correlations", out$Description)
+
+  out$Type[grepl("N_(.*)", out$Description)] <- ""
   out$Term[grepl("N_(.*)", out$Description)] <- gsub("N_(.*)", "\\1", out$Description[grepl("N_(.*)", out$Description)])
-  out$Description <- gsub("R2_(.*)", "R2 \\(\\1\\)", out$Description)
   out$Description <- gsub("_(.*)", "", out$Description)
 
+  out$Type[grepl("^X", out$Description)] <- ""
   out$Description[grepl("^X", out$Description)] <- NA
   out$Component[out$Component == ""] <- NA
   out$Term[out$Term == ""] <- NA
 
-  out[c("Description", "Component", "Term", "Value")]
+  out[c("Description", "Component", "Type", "Term", "Value")]
 }