+++ R/select_parameters.R
@@ -1,10 +1,9 @@
 #' Automated selection of model parameters
 #'
-#' This function performs an automated selection of the 'best' parameters, updating and returning the "best" model. For frequentist simple GLMs, it performs an AIC-based stepwise selection.
+#' This function performs an automated selection of the 'best' parameters, updating and returning the "best" model. For frequentist simple GLMs, it performs an AIC-based stepwise selection. For Bayesian models, it uses the \code{projpred} package.
 #'
 #' @param model A statistical model.
 #' @param ... Arguments passed to or from other methods.
-#'
 #'
 #' @examples
 #' model <- lm(mpg ~ ., data = mtcars)
@@ -32,7 +31,6 @@
 #' }
 #' }
 #' @return The model refitted with optimal number of parameters.
-#'
 #' @export
 select_parameters <- function(model, ...) {
   UseMethod("select_parameters")